requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.19
Werkzeug==0.9.4
argparse==1.2.1
itsdangerous==0.23
pyserial==2.7
wsgiref==0.1.2
